server/requirements.txt
Flask
Flask-SQLAlchemy
pydantic
psycopg2-binary
python-dotenv
flask_session